[refactor] Modify Packaging
