+++ app/src/main/res/layout/fragment_review.xml
@@ -6,7 +6,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".review.ReviewFragment">
+        tools:context=".ui.review.ReviewFragment">
 
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/rcv_review"
