+++ app/src/main/res/layout/fragment_search.xml
@@ -7,13 +7,13 @@
 
         <variable
             name="fragment"
-            type="com.example.moviereview.search.SearchFragment" />
+            type="com.example.moviereview.ui.search.SearchFragment" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".search.SearchFragment">
+        tools:context=".ui.search.SearchFragment">
 
         <EditText
             android:id="@+id/editText_search"
